+++ R/verb-copy-to.R
@@ -120,30 +120,40 @@ db_copy_to.DBIConnection <- function(con, table, values,
 
   with_transaction(con, in_transaction, {
     # Only remove if it exists; returns NA for MySQL
-    if (overwrite && !is_false(db_has_table(con, table))) {
+    if (overwrite && !is_false(dbExistsTable(con, table))) {
       db_drop_table(con, table, force = TRUE)
     }
 
     table <- db_write_table(con, table, types = types, values = values, temporary = temporary)
-    db_create_indexes(con, table, unique_indexes, unique = TRUE)
-    db_create_indexes(con, table, indexes, unique = FALSE)
+    create_indexes(con, table, unique_indexes, unique = TRUE)
+    create_indexes(con, table, indexes, unique = FALSE)
     if (analyze) db_analyze(con, table)
   })
 
   table
 }
 
+create_indexes <- function(con, table, indexes = NULL, unique = FALSE, ...) {
+  if (is.null(indexes)) return()
+  assert_that(is.list(indexes))
+
+  for (index in indexes) {
+    db_create_index(con, table, index, unique = unique, ...)
+  }
+}
+
+
 # Don't use `tryCatch()` because it messes with the callstack
 with_transaction <- function(con, in_transaction, code) {
   if (in_transaction) {
-    db_begin(con)
-    on.exit(db_rollback(con))
+    dbBegin(con)
+    on.exit(dbRollback(con))
   }
 
   code
 
   if (in_transaction) {
     on.exit()
-    db_commit(con)
+    dbCommit(con)
   }
 }