+++ tests/testthat/test-backend-oracle.R
@@ -17,3 +17,13 @@ test_that("queries translate correctly", {
   mf <- lazy_frame(x = 1, con = simulate_oracle())
   expect_snapshot(mf %>% head())
 })
+
+test_that("generates custom sql", {
+  con <- simulate_oracle()
+
+  expect_snapshot(sql_analyze(con, ident("table")))
+  expect_snapshot(sql_explain(con, sql("SELECT * FROM foo")))
+
+  expect_snapshot(sql_drop_table(con, ident("table")))
+  expect_snapshot(sql_drop_table(con, ident("table"), force = TRUE))
+})