+++ NEWS.md
@@ -1,5 +1,16 @@
 # dbplyr (development version)
 
+* A number of `db_*` generics have been replaced with SQL generation generics:
+  `sql_analyze()`, `sql_create_index()`, `sql_explain()`, `sql_drop_table()`,
+  `sql_save_query()`. This makes them easier to test is part of the process
+  of detangling dbplyr and dplyr (#284).
+  
+* A number of `db_*` generics are no longer used: `db_create_indexes()`,
+  `db_begin()`, `db_rollback()`, `db_commit()`, `db_list_tables()`,
+  `db_has_table()`.
+
+* MySQL uses standard SQL for index creation.
+
 * MS SQL translation does better a distinguishing between bit and boolean 
   (#377, #318). `if` and `ifelse` once again generate `IIF`, creating
   simpler expressions.